# api/docs/swagger.yaml
definitions:
...
  schema.APIResponse-schema_MultiBuildingEvents-schema_AstraEvent:
    properties:
      data:
        $ref: '#/definitions/schema.MultiBuildingEvents-schema_AstraEvent'
      message:
        type: string
      status:
        type: integer
    type: object
...
  schema.APIResponse-schema_SingleBuildingEvents-schema_AstraEvent:
    properties:
      data:
        $ref: '#/definitions/schema.SingleBuildingEvents-schema_AstraEvent'
      message:
        type: string
      status:
        type: integer
    type: object
...
  schema.APIResponse-string:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
  schema.AstraEvent:
    properties:
      activity_name:
        type: string
      capacity:
        type: number
      current_state:
        type: string
      end_date:
        type: string
      meeting_type:
        type: string
      not_allowed_usage_mask:
        type: number
      start_date:
        type: string
      usage_color:
        type: string
    type: object
...
  schema.MultiBuildingEvents-schema_AstraEvent:
    properties:
      buildings:
        items:
          $ref: '#/definitions/schema.SingleBuildingEvents-schema_AstraEvent'
        type: array
      date:
        type: string
    type: object
...
  schema.RoomEvents-schema_AstraEvent:
    properties:
      events:
        items:
          $ref: '#/definitions/schema.AstraEvent'
        type: array
      room:
        type: string
    type: object
...
  schema.SingleBuildingEvents-schema_AstraEvent:
    properties:
      building:
        type: string
      rooms:
        items:
          $ref: '#/definitions/schema.RoomEvents-schema_AstraEvent'
        type: array
    type: object
...
host: api.utdnebula.com
info:
  contact: {}
  description: The developer Nebula Labs API for access to pertinent UT Dallas data
  title: dev-nebula-api
  version: 1.1.0
paths:
  /astra/{date}:
    get:
      description: '"Returns AstraEvent based on the input date"'
      operationId: AstraEvents
      parameters:
      - description: date (ISO format) to retrieve astra events
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: All AstraEvents with events on the inputted date
          schema:
            $ref: '#/definitions/schema.APIResponse-schema_MultiBuildingEvents-schema_AstraEvent'
        "500":
          description: A string describing the error
          schema:
            $ref: '#/definitions/schema.APIResponse-string'
  /astra/{date}/{building}:
    get:
      description: '"Returns AstraEvent based on the input date and building name"'
      operationId: AstraEventsByBuilding
      parameters:
      - description: date (ISO format) to retrieve astra events
        in: path
        name: date
        required: true
        type: string
      - description: building abbreviation of event locations
        in: path
        name: building
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: All sections with meetings on the specified date in the specified
            building
          schema:
            $ref: '#/definitions/schema.APIResponse-schema_SingleBuildingEvents-schema_AstraEvent'
        "404":
          description: A string describing the error
          schema:
            $ref: '#/definitions/schema.APIResponse-string'
        "500":
          description: A string describing the error
          schema:
            $ref: '#/definitions/schema.APIResponse-string'
  /astra/{date}/{building}/{room}:
    get:
      description: '"Returns AstraEvent based on the input date building name and
        room number"'
      operationId: AstraEventsByBuildingandRoom
      parameters:
      - description: date (ISO format) to retrieve astra events
        in: path
        name: date
        required: true
        type: string
      - description: building abbreviation of event locations
        in: path
        name: building
        required: true
        type: string
      - description: room number for event
        in: path
        name: room
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: All sections with meetings on the specified date in the specified
            building
          schema:
            $ref: '#/definitions/schema.APIResponse-schema_SingleBuildingEvents-schema_AstraEvent'
        "404":
          description: A string describing the error
          schema:
            $ref: '#/definitions/schema.APIResponse-string'
        "500":
          description: A string describing the error
          schema:
            $ref: '#/definitions/schema.APIResponse-string'
...
schemes:
- https
- http
security:
- api_key: []
securityDefinitions:
  api_key:
    in: header
    name: x-api-key
    type: apiKey
swagger: "2.0"
...